+++ wormhole-connect/src/routes/utils.ts
@@ -122,3 +122,13 @@ export const isIlliquidDestToken = (
 export const isNttRoute = (route?: Route) => {
   return route === Route.NttManual || route === Route.NttRelay;
 };
+
+export const estimateAverageGasFee = async (
+  chain: ChainName | ChainId,
+  gasLimit: BigNumberish,
+): Promise<BigNumber> => {
+  const provider = config.wh.mustGetProvider(chain);
+  const gasPrice = await provider.getGasPrice();
+  // This is a naive estimate 30% higher than what the oracle says
+  return gasPrice.mul(gasLimit).mul(130).div(100);
+};
