+++ wormhole-connect/src/utils/gas.ts
@@ -78,13 +78,3 @@ export const estimateClaimGas = async (
   if (!gas) throw new Error('could not estimate send gas');
   return formatGasFee(destChain, gas);
 };
-
-export const estimateAverageGasFee = async (
-  chain: ChainName | ChainId,
-  gasLimit: BigNumberish,
-): Promise<BigNumber> => {
-  const provider = config.wh.mustGetProvider(chain);
-  const gasPrice = await provider.getGasPrice();
-  // This is a naive estimate 30% higher than what the oracle says
-  return gasPrice.mul(gasLimit).mul(130).div(100);
-};
